requirements.txt
Flask==1.1.2
oauth2client==4.1.3
opencv-python==4.3.0.36
Werkzeug==1.0.1

numpy~=1.19.2
gspread~=3.7.0
scikit-learn~=0.23.2
google-cloud-vision==2.3.2